+++ packages/components-js/tests/vrt/specs/states/checkbox.vrt.ts
@@ -47,13 +47,13 @@ const scenario = async (page: Page, theme: Theme, scheme?: PrefersColorScheme): 
   });
 
   await forceHoverState(page, '.hover p-checkbox:not(.force-label) >>> input');
-  await forceHoverState(page, '.hover p-checkbox.force-label >>> label');
+  await forceHoverState(page, '.hover p-checkbox.force-label >>> .label-wrapper');
   await forceHoverState(page, '.hover p-checkbox span a');
   await forceFocusVisibleState(page, '.focus p-checkbox >>> input');
   await forceFocusVisibleState(page, '.focus p-checkbox span a');
   await forceFocusHoverState(page, '.focus-hover p-checkbox:not(.force-label) >>> input');
   await forceFocusVisibleState(page, '.focus-hover p-checkbox.force-label >>> input');
-  await forceHoverState(page, '.focus-hover p-checkbox.force-label >>> label');
+  await forceHoverState(page, '.focus-hover p-checkbox.force-label >>> .label-wrapper');
   await forceFocusHoverState(page, '.focus-hover p-checkbox span a');
 };