+++ packages/storefront/src/app/components/input-password/input-password.stories.ts
@@ -85,7 +85,7 @@ export const inputPasswordStorySlots: Story<'p-input-password'> = {
       children: [
         {
           tag: 'span',
-          properties: { slot: 'label', id: 'some-label-id' },
+          properties: { slot: 'label' },
           children: [
             'Some label with a ',
             { tag: 'a', properties: { href: 'https://designsystem.porsche.com' }, children: ['link'] },
@@ -94,7 +94,7 @@ export const inputPasswordStorySlots: Story<'p-input-password'> = {
         },
         {
           tag: 'span',
-          properties: { slot: 'description', id: 'some-description-id' },
+          properties: { slot: 'description' },
           children: [
             'Some description with a ',
             { tag: 'a', properties: { href: 'https://designsystem.porsche.com' }, children: ['link'] },
@@ -103,7 +103,7 @@ export const inputPasswordStorySlots: Story<'p-input-password'> = {
         },
         {
           tag: 'span',
-          properties: { slot: 'message', id: 'some-message-id' },
+          properties: { slot: 'message' },
           children: [
             'Some error message with a ',
             { tag: 'a', properties: { href: 'https://designsystem.porsche.com' }, children: ['link'] },